requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.6
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.6
PyMySQL==1.1.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
